Add Code - Class로 변경 작업 시작
